实现贝叶斯分类器

+++ src/main/java/com/netease/ysf/shine/classify/DataLine.java
@@ -6,6 +6,5 @@ import lombok.Data;
 public class DataLine {
 
     private int category;
-    private float[] vector;
-
+    private double[] vector;
 }
